refactor: 创建 plugin_sdk 包，重构插件系统架构
- 创建 plugin_sdk 包，将插件开发 API 独立出来
- 移动 config_types、plugin_base、service_registry、server_bridge 到 plugin_sdk
- 移动 shared_types/services 到 plugins/services
- 重构配置系统：create_widget 返回 ConfigWidgetBase
- 添加服务等待机制 wait_for_service
- 重构 history 插件为包形式
- 更新插件管理器 UI：按钮移至工具栏
- 更新 build.bat 和 plugin-dev-tutorial.md
