+++ src/plugin_sdk/config_types/base_config.py
@@ -16,6 +16,67 @@ from PyQt5.QtCore import pyqtSignal, QObject
 T = TypeVar("T")
 
 
+class ConfigWidgetBase(QWidget):
+    """
+    配置控件基类
+
+    自定义配置类型的控件必须继承此类，并实现以下内容：
+    - get_value(): 获取当前值
+    - set_value(value): 设置当前值
+    - value_change 信号: 值变化时发射（已提供默认实现）
+    """
+
+    value_change = pyqtSignal(object)
+
+    def get_value(self) -> Any:
+        """获取当前值"""
+        raise NotImplementedError("子类必须实现 get_value 方法")
+
+    def set_value(self, value: Any) -> None:
+        """设置当前值"""
+        raise NotImplementedError("子类必须实现 set_value 方法")
+
+
+class ConfigWidgetWrapper(ConfigWidgetBase):
+    """
+    配置控件包装器
+
+    将现有 Qt 控件包装为 ConfigWidgetBase，用于简化内置配置类型的实现。
+    """
+
+    def __init__(
+        self,
+        widget: QWidget,
+        getter: Callable[[], Any],
+        setter: Callable[[Any], None],
+        signal: QObject,
+        parent: QWidget | None = None,
+    ):
+        super().__init__(parent)
+        self._widget = widget
+        self._getter = getter
+        self._setter = setter
+
+        # 将控件添加到布局
+        from PyQt5.QtWidgets import QVBoxLayout
+        layout = QVBoxLayout(self)
+        layout.setContentsMargins(0, 0, 0, 0)
+        layout.addWidget(widget)
+
+        # 连接原始信号到 value_change
+        signal.connect(self._on_value_changed)
+
+    def _on_value_changed(self) -> None:
+        """内部信号变化时转发到 value_change"""
+        self.value_change.emit(self._getter())
+
+    def get_value(self) -> Any:
+        return self._getter()
+
+    def set_value(self, value: Any) -> None:
+        self._setter(value)
+
+
 @dataclass
 class BaseConfig(ABC, Generic[T]):
     """
@@ -48,15 +109,15 @@ class BaseConfig(ABC, Generic[T]):
             self.label = ""
 
     @abstractmethod
-    def create_widget(self) -> tuple[QWidget, Callable[[], T], Callable[[T], None], QObject]:
+    def create_widget(self) -> ConfigWidgetBase:
         """
         创建 PyQt 控件
 
         Returns:
-            (控件, 获取值函数, 设置值函数, 值变化信号对象)
-            
-            信号对象应该是一个有 connect 方法的 QObject（如 pyqtSignal）。
-            当值变化时，配置系统会自动连接这个信号来同步值。
+            ConfigWidgetBase 实例，必须实现:
+            - get_value(): 获取当前值
+            - set_value(value): 设置当前值
+            - value_change 信号: 值变化时发射
         """
         pass
