+++ src/plugin_sdk/config_types/bool_config.py
@@ -4,12 +4,11 @@
 
 from __future__ import annotations
 
-from typing import Any, Callable
+from typing import Any
 
 from PyQt5.QtWidgets import QCheckBox
-from PyQt5.QtCore import QObject
 
-from .base_config import BaseConfig
+from .base_config import BaseConfig, ConfigWidgetWrapper
 
 
 class BoolConfig(BaseConfig[bool]):
@@ -27,13 +26,13 @@ class BoolConfig(BaseConfig[bool]):
         """确保默认值是布尔类型"""
         self.default = bool(self.default)
 
-    def create_widget(self) -> tuple[QCheckBox, Callable[[], bool], Callable[[bool], None], QObject]:
-        """创建 QCheckBox 控件，返回 (控件, getter, setter, 信号)"""
+    def create_widget(self) -> ConfigWidgetWrapper:
+        """创建 QCheckBox 控件"""
         widget = QCheckBox()
         widget.setChecked(self.default)
         if self.description:
             widget.setToolTip(self.description)
-        return widget, widget.isChecked, widget.setChecked, widget.stateChanged
+        return ConfigWidgetWrapper(widget, widget.isChecked, widget.setChecked, widget.stateChanged)
 
     def to_storage(self, value: bool) -> bool:
         """转换为存储格式"""
