+++ src/plugin_sdk/config_types/text_config.py
@@ -5,12 +5,11 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Any, Callable
+from typing import Any
 
-from PyQt5.QtCore import QObject
 from PyQt5.QtWidgets import QLineEdit
 
-from .base_config import BaseConfig
+from .base_config import BaseConfig, ConfigWidgetWrapper
 
 
 @dataclass
@@ -40,8 +39,8 @@ class TextConfig(BaseConfig[str]):
         """确保默认值是字符串类型"""
         self.default = str(self.default)
 
-    def create_widget(self) -> tuple[QLineEdit, Callable[[], str], Callable[[str], None], QObject]:
-        """创建 QLineEdit 控件，返回 (控件, getter, setter, 信号)"""
+    def create_widget(self) -> ConfigWidgetWrapper:
+        """创建 QLineEdit 控件"""
         widget = QLineEdit()
         widget.setText(str(self.default))
 
@@ -54,7 +53,7 @@ class TextConfig(BaseConfig[str]):
         if self.description:
             widget.setToolTip(self.description)
 
-        return widget, widget.text, widget.setText, widget.textChanged
+        return ConfigWidgetWrapper(widget, widget.text, widget.setText, widget.textChanged)
 
     def to_storage(self, value: str) -> str:
         """转换为存储格式"""
