+++ src/plugins/__init__.py
@@ -1,5 +1,5 @@
 """
-示例插件目录
+插件包
 
-将插件放在此目录下，插件管理器会自动发现和加载。
+此目录包含所有插件模块和服务定义。
 """
